+++ src/lib.rs
@@ -1,4 +1,4 @@
-#![doc = "Peripheral access API for STM32F30X microcontrollers (generated using svd2rust v0.13.0)\n\nYou can find an overview of the API [here].\n\n[here]: https://docs.rs/svd2rust/0.13.0/svd2rust/#peripheral-api"]
+#![doc = "Peripheral access API for STM32F30X microcontrollers (generated using svd2rust v0.13.1)\n\nYou can find an overview of the API [here].\n\n[here]: https://docs.rs/svd2rust/0.13.1/svd2rust/#peripheral-api"]
 #![deny(missing_docs)]
 #![deny(warnings)]
 #![allow(non_camel_case_types)]
@@ -210,53 +210,6 @@ pub static __INTERRUPTS: [Vector; 82] = [
     Vector { _reserved: 0 },
     Vector { _handler: FPU },
 ];
-#[doc = r" Macro to override a device specific interrupt handler"]
-#[doc = r""]
-#[doc = r" # Syntax"]
-#[doc = r""]
-#[doc = r" ``` ignore"]
-#[doc = r" interrupt!("]
-#[doc = r"     // Name of the interrupt"]
-#[doc = r"     $Name:ident,"]
-#[doc = r""]
-#[doc = r"     // Path to the interrupt handler (a function)"]
-#[doc = r"     $handler:path,"]
-#[doc = r""]
-#[doc = r"     // Optional, state preserved across invocations of the handler"]
-#[doc = r"     state: $State:ty = $initial_state:expr,"]
-#[doc = r" );"]
-#[doc = r" ```"]
-#[doc = r""]
-#[doc = r" Where `$Name` must match the name of one of the variants of the `Interrupt`"]
-#[doc = r" enum."]
-#[doc = r""]
-#[doc = r" The handler must have signature `fn()` is no state was associated to it;"]
-#[doc = r" otherwise its signature must be `fn(&mut $State)`."]
-#[cfg(feature = "rt")]
-#[macro_export]
-macro_rules! interrupt {
-    ($Name:ident, $handler:path,state: $State:ty = $initial_state:expr) => {
-        #[allow(unsafe_code)]
-        #[deny(private_no_mangle_fns)]
-        #[no_mangle]
-        pub unsafe extern "C" fn $Name() {
-            static mut STATE: $State = $initial_state;
-            let _ = $crate::Interrupt::$Name;
-            let f: fn(&mut $State) = $handler;
-            f(&mut STATE)
-        }
-    };
-    ($Name:ident, $handler:path) => {
-        #[allow(unsafe_code)]
-        #[deny(private_no_mangle_fns)]
-        #[no_mangle]
-        pub unsafe extern "C" fn $Name() {
-            let _ = $crate::Interrupt::$Name;
-            let f: fn() = $handler;
-            f()
-        }
-    };
-}
 #[doc = r" Enumeration of all the interrupts"]
 pub enum Interrupt {
     #[doc = "0 - Window Watchdog interrupt"]
@@ -465,10 +418,12 @@ unsafe impl ::bare_metal::Nr for Interrupt {
         }
     }
 }
-#[doc(hidden)]
-pub mod interrupt;
+#[cfg(feature = "rt")]
+pub use self::Interrupt as interrupt;
 pub use cortex_m::peripheral::Peripherals as CorePeripherals;
 pub use cortex_m::peripheral::{CBP, CPUID, DCB, DWT, FPB, FPU, ITM, MPU, NVIC, SCB, SYST, TPIU};
+#[cfg(feature = "rt")]
+pub use cortex_m_rt::interrupt;
 #[doc = "General-purpose I/Os"]
 pub struct GPIOA {
     _marker: PhantomData<*const ()>,
@@ -1417,6 +1372,7 @@ impl Deref for OPAMP {
 }
 #[doc = "Operational amplifier"]
 pub mod opamp;
+#[allow(renamed_and_removed_lints)]
 #[allow(private_no_mangle_statics)]
 #[no_mangle]
 static mut DEVICE_PERIPHERALS: bool = false;